*Fixed compiler warnings.

+++ otserv/trunk/creature.cpp
@@ -912,7 +912,7 @@ DeathList Creature::getKillers(int assist_count)
 			std::sort(list.begin() + 1, list.end(), DeathLessThan());
 	}
 
-	if(list.size() > assist_count + 1)
+	if(list.size() > (uint32_t)assist_count + 1)
 		// Shrink list to assist_count
 		list.resize(assist_count + 1, DeathEntry(NULL, -1));
